Formulario para agregar informacion de zonas

+++ Source/Backend/database/migrations/2017_09_22_013424_create_areas_table.php
@@ -15,7 +15,7 @@ class CreateAreasTable extends Migration
     {
         Schema::create('areas', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('measure');
+            $table->decimal('measure');
             $table->char('unit', 10);
             $table->integer('id_zone')->unsigned();
             $table->foreign('id_zone')
